% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deviant.R
\name{most_deviant}
\alias{most_deviant}
\title{Most deviant case}
\usage{
most_deviant(lmobject)
}
\arguments{
\item{lmobject}{Object generated with \code{\link[stats]{lm}}}
}
\value{
The most deviant case having the largest absolute
residual of all cases. Originally proposed by
Originally proposed by Seawright and Gerring
(\href{https://journals.sagepub.com/doi/10.1177/1065912907313077}{2008}).
}
\description{
Most deviant case
}
\examples{
df <- lm(mpg ~ disp + wt, data = mtcars)
most_deviant(df)
}
