% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deviant.R
\name{most_overpredicted}
\alias{most_overpredicted}
\title{Most overpredicted case}
\usage{
most_overpredicted(lmobject)
}
\arguments{
\item{lmobject}{Object generated with \code{\link[stats]{lm}}}
}
\value{
The most overpredicted case with the largest negative residual
(the most negative residual).
}
\description{
Most overpredicted case
}
\examples{
df <- lm(mpg ~ disp + wt, data = mtcars)
most_overpredicted(df)
}
