% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deviant.R
\name{most_underpredicted}
\alias{most_underpredicted}
\title{Most underpredicted case}
\usage{
most_underpredicted(lmobject)
}
\arguments{
\item{lmobject}{Object generated with \code{\link[stats]{lm}}}
}
\value{
The most underpredicted case with the largest positive residual
(the most positive residual).
}
\description{
Most underpredicted case
}
\examples{
df <- lm(mpg ~ disp + wt, data = mtcars)
most_underpredicted(df)
}
