% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway.R
\name{pathway}
\alias{pathway}
\title{Pathway case}
\usage{
pathway(full.model, reduced.model)
}
\arguments{
\item{full.model}{Full model including covariate of interest}

\item{reduced.model}{Reduced model excluding covariate of interest}
}
\value{
A dataframe with pathway case residuals.
}
\description{
Pathway case
}
\examples{
df_full <- lm(mpg ~ disp + wt, data = mtcars)
df_reduced <- lm(mpg ~ wt, data = mtcars)
pathway(df_full, df_reduced)

}
