% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/most_typical.R
\name{most_typical}
\alias{most_typical}
\title{Most typical case}
\usage{
most_typical(lmobject)
}
\arguments{
\item{lmobject}{Object generated with \code{\link[stats]{lm}}}
}
\value{
The most typical case having the smallest absolute
residual of all cases.
}
\description{
Most typical case
}
\examples{
df <- lm(mpg ~ disp + wt, data = mtcars)
most_typical(df)
}
