% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway.R
\name{pathway}
\alias{pathway}
\title{Pathway case}
\usage{
pathway(full.model, reduced.model)
}
\arguments{
\item{full.model}{Full model including covariate of interest}

\item{reduced.model}{Reduced model excluding covariate of interest}
}
\value{
A dataframe with pathway case residuals (pathway values) for
both definitions of pathway values.
}
\description{
Calculation of difference between residuals of full model and reduced
model lacking the focal variable. The larger the difference, the more
a case qualifies as a pathway case suitable for the analysis of
mechanisms.
}
\details{
The plain difference follows the approach developed by Weller and Barnes
(Weller, Nicholas and Jeb Barnes (2014): \emph{Finding Pathways: Mixed-Method
Research for Studying Causal Mechanisms.} Cambridge: Cambridge University
Press. \url{https://doi.org/10.1017/CBO9781139644501}).

The calculation of the absolute residual if the reduced-model
residual is larger than the full-model residual follows the original
proposal by Gerring (Gerring, John (2007): Is There a (Viable)
Crucial-Case Method? \emph{Comparative Political Studies} 40 (3): 231-253.
\url{https://journals.sagepub.com/doi/10.1177/0010414006290784})
}
\examples{
df_full <- lm(mpg ~ disp + wt, data = mtcars)
df_reduced <- lm(mpg ~ wt, data = mtcars)
pathway(df_full, df_reduced)

}
