% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway.R
\name{pathway_xvr}
\alias{pathway_xvr}
\title{Plot of residuals against pathway variable}
\usage{
pathway_xvr(
  full.model,
  reduced.model,
  pathway.var = "variable",
  pathway.type = "residual"
)
}
\arguments{
\item{full.model}{Full model including covariate of interest (=pathway variable)}

\item{reduced.model}{Reduced model excluding covariate of interest}

\item{pathway.var}{Pathway variable dropped from full model}

\item{pathway.type}{Type of pathway values. \code{pathway.wb} are
pathway values proposed by Weller and Barnes. \code{pathway.g} are values
as calculated by Gerring.}
}
\value{
A plot of the chosen type of pathway values against the pathway variable
created with \code{\link{ggplot2}}.
}
\description{
Plot of residuals against pathway variable
}
\examples{
df_full <- lm(mpg ~ disp + wt, data = mtcars)
df_reduced <- lm(mpg ~ wt, data = mtcars)
pathway_xvr(full.model, reduced.model, pathway.var = "disp",
pathway.type = "pathway.wb")

}
