% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{predint}
\alias{predint}
\title{Classification of cases as typical and deviant using the prediction
interval.}
\usage{
predint(lmobject, piwidth = 0.95)
}
\arguments{
\item{lmobject}{Object generated with \code{\link[stats]{lm}}}

\item{piwidth}{Width of the prediction interval.}
}
\value{
A dataframe with observed outcome, fitted outcome values,
upper and lower bound of the prediction interval and classification
of cases as typical or deviant. Originally proposed by Rohlfing and
Starke (\url{https://doi.org/10.1111/spsr.12052}).
}
\description{
Classification of cases as typical and deviant using the prediction
interval.
}
\examples{
df <- lm(mpg ~ disp + wt, data = mtcars)
predint(df, piwidth = 0.9)

}
